blog/flask-blog/requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
Flask==1.0.2
Flask-Session==0.3.1
gevent==1.3.6
greenlet==0.4.14
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
Pillow==5.2.0
pycparser==2.18
pycryptodome==3.6.6
PyMySQL==0.9.2
redis==2.10.6
requests==2.19.1
six==1.11.0
UNKNOWN==0.0.0
urllib3==1.23
Werkzeug==0.14.1
